Enemy Implemented & Player modifications Made
I've implemented the enemy

They have health, and can die when it goes to zero.

The default bullets do 1 damage,
The enemies have 1 health.
They give 5 xp when killed.
They play an explosion effect when dying that needs to be replaced.

TODO:
The enemy is only missing an attack, probably just going to copy the code for the player's shooting for their attack.

I still need to implement them spawning at the edges of the world or just outside the camera and in waves.

Shouldn't be terribly difficult.

CHANGES MADE FOR THE PLAYER:

I changed it so WASD and Arrows can be used for input, and mouse is used for aiming.

I also changed quite a bit of how XP was setup so that I could implement the level up code. It just checks for a level up every time you gain xp, James will be able to go in and fine tune it to his liking, I just put a temp formula in there for now.

There's also a temp GUI draw in a obj_game object I made just so we can see the current level and the current XP, it can be removed in the future. I should also add the player's health to it later.

TODO:
The player needs a death like how the enemy currently has one.

I should also add a method that is called when you actually level up so that Ryan can use that to setup the UI for picking your upgrades.

I also changed the way movement was handled so that I could make the player's diagonal speed the same has moving on one of the horizontal or vertical axes.

I AM NOT DONE CODING YET, DO NOT MODIFY.