Implemented Gun Animation
I implemented a firing animation for the gun and a sprite for the player bullet.

Now I just need to make a copy of the player bullet that is recolored to magenta or neon purple for the enemy.

Then I'll just go back to working on the player and the enemy designs.