Implemented Debug Controls & Burst Fire
I implemented some controls so you can test upgrading the various abilities you have.

I also added a "burst" upgrade because there wasn't enough variety in weapon upgrades imo. But, we don't have to include it, it was just really easy to implement.

The obj_temp_gui can be deleted and all the debug menu stuff won't be in the game so no need to worry about tracking it all down.

It's just there to show y'all how to access it.

TODO:

LD needs to finish working on enemy designs and player.

Ryan can now work on UI.