Update xp UI and Upgrade UI
The xp bar now shows progress from the current level to the next level.

The upgrade UI now has categories so that the first slot is always a weapon upgrade and the 2nd and third slot are the "other" category. We can make it so that 2 of the slots are always weapons and the 3rd is "other".

TODO:
Focus on the scaling of the enemies and how waves scale, there also seems to be an issue where at wave 10 killing all enemies doesn't progress to the next wave. I need to figure out why.

I also need to implement the art for the player and the enemies.