Implemented Player Animations & bug fix
I fixed an issue where some objects were destroyed by a bullet right when the game paused so the object it found became null so I just added a null check in the obj_time_stop code.

I implemented the idle and walking sprites for the player so they now play walking when moving and idle when stationary.

I need to move where the gun visual is displayed as it currently looks odd. When the player is facing up the gun should be rendered under their sprite and when the player is facing down the gun should be rendered over their sprite. When facing left or right the gun should be drawn on top.