Updated Gun Code
I updated the code for drawing the gun and spawning the bullet so that we can apply an offset to the gun's position. The spawning of the bullet now occurs exactly at the end of the gun barrel. That way it's more consistent than before.

I'm going to work on making the enemy designs next and making a couple of diverse enemies. My main focus is to make 1 idle sprite and 1 movement animation and not have the enemy rotate at all. I'm thinking one of the enemies could be a squid/octopus that floats by moving their tentacles up then down like how they swim. They'd over and they'd shoot 8 bullets in all directions. That way they'd be a more diverse enemy. Then the base enemy could be some kind of soldier or another space invaders-style sprite like james suggested.