Testing bug fixes for the spawner.
The way that the pathfinding works causes enemies to be "stuck" if they spawn on top of each other. I have implemented a crude method to prevent enemies from getting stuck on each other. The spawner will spawn them separately and on top of that if they collide with another enemy they will move offset from them.

The game gets stuck on wave 13 or 12 because it seems some enemies get stuck on each other EVEN though they have the collision check.

I need to take a break.