Fixed Spawning System.
I have completely fixed how we spawn waves so that the enemies spawned are staggered and we don't spawn all the enemies at once. This fixes the issue where enemies would be spawned on top of each other and the game would not be able to progress as those enemies would be stuck outside the room.

TODO:
LD wants to play the game a ton and work on the balancing as it could use a little more work.

LD is also going to go in and do a general cleanup of how the healthbar & xpbar are drawn as it looks kind of messy.