Spawner Bug fix
I just fixed the final issue with the spawner where it was always 5 kills to reach the next wave instead of a larger number.

This was because I was saying kills to spawn next wave = cur_wave_size instead of adding to it.

The balancing is quite good for an "easy" game.

Right now the enemies stats do not scale with the wave.

it took me around 30 minutes to reach wave 50 and win.

I also only reached the minimum speed for firing interval (10) at level 45. At that speed you're basically invincible.

Getting to piercing 4 and a burst 3 and a bullets per burst 3 is basically the minimum you can beat the game at.

I'll have to figure out how to scale the enemies to make them more difficult, I may also try changing how fast the firing interval goes down and the chances of piercing showing up.

We'll see.