Added global vars for Kills and Waves.
I made it so that the kills and waves are global so they can be referenced after the main scene ends.

TODO:
Make the regen start worse than it currently is, change up how that stuff works. I'm too tired to think about it but I also couldn't sleep because I realized how to fix an issue.

I still need to work on balancing the enemies and make them scale with wave. The player absorbs damage far too easily right now.