Implemented visual damage feedback.
I implemented particle systems for when the enemies die. They explode into little cubes like when getting derezzed in TRON.

I also implemented "Flare effects" which play when a player or an enemy takes damage.

I have balanced the game pretty well but it feels a little to difficult.

I think I'm going to add another enemy that's melee-based just because we really need it for balancing purposes. I want to make the tank enemy start showing up in later rounds and not be there at the beginning. I also want to make the tank enemies a bit weaker but that's not important right now.

I think I need to convert the game to be more like vampire survivors so that there are multiple weapons and the game feels more diverse in it's weaponry and it'll make balancing much easier.

I don't want the game to feel like you're just upgrading one thing and I think as an extra along side the additional characters we should add more weapons.

I want davis to play the game tho so we can get permission to move on to extra stuff as a group.

I also made it so that the temp_gui stuff I made is now a cheat screen that can be opened by pressing "O".