Began Implementing "Juice"
I started implementing some effects for the UI for when you level up/when you gain xp. The xp bar actually animates to fill instead of jumping to fill.

This isn't a complete implementation of it because I want to commit before I implement the same thing for the health bar and possibly break my code.

I also need to make it so that the xp bar animates to be full and then rolls over and restarts when you level up.

The other thing I did was make the LVL. text do a bouncy animation when you level up so there's a little more visual indication of it.