Melee Enemy & Spawner Changes
I implemented the Melee enemy (but not their art) and I changed the spawner so spawning is more staggered than before. The enemies are more spread out now.

Melee enemies spawn until wave 15 when tank enemies have a chance to spawn instead of a melee enemy.

I need to make the enemies health and damage properly scale with wave.