Fixed some issues with enemy heatlh.
I fixed an issue with enemy health where everything was using the health variable that was global in default by gamemaker so I changed them to use an individual variable. I cannot believe I forgot, that was in one of the workshops.

TODO: Replace enemy health scaling with the same formula the player uses but make their health grow faster.

Replace the xp the enemies give so that it scales with their level.

Add sprite art for the melee enemies.