Implement squid Sprite & tweak balancing
I implemented the new squid sprite and I tweaked balancing for the game a little. I'm going to work on tweaking the xp values rewarded by enemies and I'm also going to work on the enemy health scaling so the tanks are "tankier" than the melee enemies.

I also need to implement the pause screen that ryan made for the UI.

Right now the game is fairly well balanced and I just played and died at wave 38.

I need xp to scale but I also need enemy health and damage to scale.

I also want to make it so that the enemies have a larger cool down between spawner bursts of enemies within a wave. That way all the enemies don't spawn at once in a circle around the player and instead spawn more staggered over time around the player so that it's multiple circles for one wave and feels more like the enemies are approaching over time instead of one big burst per wave.