Update Spawner
I implemented the staggering code for spawning waves so that there are bursts of enemies that spawn and have a cooldown between bursts within a single wave. This added difficulty as the enemies spawning is not a basic pattern and instead is much less predictable. This gives me what I wanted.