Upped the difficulty a lot.
I increased the difficulty a lot by making enemy health scale with wave, I think it's really important I do this before I implement scaling to increase xp.

I also made it so that the player's base damage increases as they increase in level. This isn't an upgrade as we aren't super focused on individual weapon upgrades. If we did decide to do weapon upgrades I'd probably have to change it so that the damage can be upgraded.

We should also do percentage based upgrades instead of what it looks like now.

So, I'm going to commit this then work on implementing xp scaling.

Actually, I have class in a bit so I gotta go to that first.

But after I do xp scaling I'll finish implementing the pause menu.