Implement XP scaling for enemies
The xp rewarded by enemies on kill now scales with wave. It doesn't increase extremely often and I need to work on the formula for it but I'm uploading this as my class is about to start.

I also need to change the regen upgrade to increase the amount rewarded by regen relative to your health as it currently doesn't scale properly.