Update obj_xp_bar
I moved all the code for drawing and animating the xp bar into the object for the xp bar so that it is better abstracted, it's also for readability as I try to figure out why the rollover for level doesn't 100% work.

I'm going to do that for the health bar as well, just not at the moment. I'm instead going to make sure the pause menu is set up before I call it a night.