Pause menu fully implemented
I have the pause menu fully implemented now such that you can pause the game whenever and the buttons on the menu work. This also means pausing during the upgrade selection opens the menu and the game remains frozen, and closing the menu during the upgrade selection still leaves the game frozen as the upgrade menu is freezing the game.

The sliders for audio work, I just need to add labels above them.