Update Health Regen & add controls
I added the health regen so it's a little more useful to upgrade. It now makes the game balanced enough such that you could beat the game given the right upgrades and strategy, or lose without them.

I also added the controls to the rules screen because we didn't have those.