Bug fixes & Game speed cheat
I still am not a huge fan of the changes to balancing due to changing the enemy sizes and speeds early game. I may need to change the speed back but not the size because it really affected the game play.

The next thing is that I figured out how to set the current room's speed so that the game fps is higher and the game runs "faster". This is how one could also implement the time-slow ability that james wanted to do.

The bugs I fixed had to do with the global variables not getting reset when exiting the main room.