Bug Fixes
Fixed text on game over screen (it is now centered properly and displays the correct wave number)

Made it so that if you level up multiple times before selecting an upgrade the game waits to let the user select an upgrade before prompting to select another upgrade for the additional levels they gained.

I'm doing this so I can figure out how to fix the xp rollover visuals.